requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
asttokens==3.0.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
Django==4.2.10
django-cors-headers==4.3.0
django-sslserver==0.22
executing==2.2.0
frozenlist==1.8.0
greenlet==3.3.0
groq==0.37.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.12
langchain-core==0.3.63
langchain-groq==0.2.1
langchain-text-splitters==0.3.8
langsmith==0.2.11
matplotlib-inline==0.1.7
multidict==6.7.0
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.11.4
packaging==24.2
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.4.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
pywin32==310
PyYAML==6.0.3
pyzmq==26.4.0
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.4
stack-data==0.6.3
tenacity==9.1.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
wcwidth==0.2.13
whitenoise==6.5.0
yarl==1.22.0
